Move json model to separate crate
This commit reorganizes the project structure by moving `crossmap.rs` and `json.rs` into a new `nativedocgen` crate. It also creates a new `nativedocgen-model` crate to hold the data models used in the project. The dependencies have been updated accordingly, with the addition of the new crates as well as changes to existing dependencies.